Move env_creator to rail_env_utils.
